peip/requirements.txt
pydriller
fastapi
uvicorn
openai
python-dotenv
supabase
pandas
radon
requests
